Added Axios
I added axios dependency to my project as I could not make it work with fetch. I managed to make an axios request based on a click event and it does fetch the data to the console.
I also added dotenv to the project for the API key and swapped all the values in my files. I also added the .env to gitignore