Added the allCountries part
I managed to get the allCountries part working. It now shows all countries and holidays on click. They are divided in their own boxes to look nice.
The whole code is working as intended and fetched the data for the specific country on click.
I added footer and styled it
I added some styling to the overall page too