Added a search bar
I finished with the app and it works fine but I decided to add a search bar for the countries page as they are a lot and it will be difficult for the user to find them.
Decided to go with fuse,js. I downloaded and started fixing my code as it did not want to work. I changed my setFetch method with Axios and managed to render all the data.
Next stage will be to add the code in the fuse.js in order to make it loop through the countries array and show only the ones that have the typed text in their name.